requirements.txt
asgiref==3.11.1
celery==5.3.4
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
django-redis==5.4.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
flower==2.0.1
gunicorn==22.0.0
idna==3.11
Levenshtein==0.27.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-Levenshtein==0.27.3
RapidFuzz==3.14.5
redis==5.0.1
requests==2.33.1
spacy==3.8.14
sqlparse==0.5.5
urllib3==2.6.3
